pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "delta-audit"
version = "0.1.0"
description = "A lightweight Δ-Attribution suite for auditing model updates."
authors = [{name="Arshia Hemmat"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.20",
    "pandas>=1.2",
    "scikit-learn>=1.0",
    "matplotlib>=3.4",
    "scipy>=1.7",
    "joblib>=1.1",
    "pyyaml>=6.0",
    "seaborn>=0.11"
]

[project.scripts]
delta-audit = "delta_audit.cli:main"

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "black>=23.0",
    "pytest>=7.0",
    "pre-commit>=3.0",
    "codespell>=2.2"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B", "W", "C90"]
ignore = ["E501", "W503"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"] 